Revert "Don't block /get_watch"
