+++ app/src/main/java/app/revanced/integrations/youtube/patches/spoof/SpoofClientPatch.java
@@ -2,16 +2,21 @@ package app.revanced.integrations.youtube.patches.spoof;
 
 import android.net.Uri;
 
-import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
+
+import org.chromium.net.UrlRequest;
 
 import java.nio.ByteBuffer;
 import java.util.Map;
+import java.util.Objects;
+import java.util.concurrent.ExecutionException;
+import java.util.concurrent.Future;
+import java.util.concurrent.TimeUnit;
+import java.util.concurrent.TimeoutException;
 
 import app.revanced.integrations.shared.Logger;
 import app.revanced.integrations.shared.Utils;
-import app.revanced.integrations.youtube.patches.VideoInformation;
-import app.revanced.integrations.youtube.patches.spoof.requests.StreamingDataRequest;
+import app.revanced.integrations.youtube.patches.spoof.requests.StreamingDataRequester;
 import app.revanced.integrations.youtube.settings.Settings;
 
 @SuppressWarnings("unused")
@@ -22,8 +27,34 @@ public class SpoofClientPatch {
      * Any unreachable ip address.  Used to intentionally fail requests.
      */
     private static final String UNREACHABLE_HOST_URI_STRING = "https://127.0.0.0";
+    private static final Uri UNREACHABLE_HOST_URI = Uri.parse(UNREACHABLE_HOST_URI_STRING);
 
-    private static volatile Map<String, String> fetchHeaders;
+    private static volatile Future<ByteBuffer> currentVideoStream;
+
+    /**
+     * Injection point.
+     * Blocks /get_watch requests by returning an unreachable URI.
+     *
+     * @param playerRequestUri The URI of the player request.
+     * @return An unreachable URI if the request is a /get_watch request, otherwise the original URI.
+     */
+    public static Uri blockGetWatchRequest(Uri playerRequestUri) {
+        if (SPOOF_CLIENT) {
+            try {
+                String path = playerRequestUri.getPath();
+
+                if (path != null && path.contains("get_watch")) {
+                    Logger.printDebug(() -> "Blocking 'get_watch' by returning unreachable uri");
+
+                    return UNREACHABLE_HOST_URI;
+                }
+            } catch (Exception ex) {
+                Logger.printException(() -> "blockGetWatchRequest failure", ex);
+            }
+        }
+
+        return playerRequestUri;
+    }
 
     /**
      * Injection point.
@@ -59,45 +90,28 @@ public class SpoofClientPatch {
     /**
      * Injection point.
      */
-    public static void setFetchHeaders(String url, Map<String, String> headers) {
+    public static UrlRequest buildRequest(UrlRequest.Builder builder, String url,
+                                          Map<String, String> playerHeaders) {
         if (SPOOF_CLIENT) {
             try {
                 Uri uri = Uri.parse(url);
                 String path = uri.getPath();
-                if (path != null && path.contains("browse")) {
-                    fetchHeaders = headers;
+                if (path != null && path.contains("player") && !path.contains("heartbeat")) {
+                    String videoId = Objects.requireNonNull(uri.getQueryParameter("id"));
+                    currentVideoStream = StreamingDataRequester.fetch(videoId, playerHeaders);
                 }
             } catch (Exception ex) {
-                Logger.printException(() -> "setFetchHeaders failure", ex);
+                Logger.printException(() -> "buildRequest failure", ex);
             }
         }
+
+        return builder.build();
     }
 
     /**
      * Injection point.
-     */
-    public static void fetchStreamingData(@NonNull String videoId, boolean isShortAndOpeningOrPlaying) {
-        if (SPOOF_CLIENT) {
-            try {
-                final boolean videoIdIsShort = VideoInformation.lastPlayerResponseIsShort();
-                // Shorts shelf in home and subscription feed causes player response hook to be called,
-                // and the 'is opening/playing' parameter will be false.
-                // This hook will be called again when the Short is actually opened.
-                if (videoIdIsShort && !isShortAndOpeningOrPlaying) {
-                    return;
-                }
-
-                StreamingDataRequest.fetchRequestIfNeeded(videoId, fetchHeaders);
-            } catch (Exception ex) {
-                Logger.printException(() -> "fetchStreamingData failure", ex);
-            }
-	    }
-	}
-
-    /**
-     * Injection point.
      * Fix playback by replace the streaming data.
-     * Called after {@link #setFetchHeaders(String, Map)} .
+     * Called after {@link #buildRequest(UrlRequest.Builder, String, Map)}.
      */
     @Nullable
     public static ByteBuffer getStreamingData(String videoId) {
@@ -105,17 +119,23 @@ public class SpoofClientPatch {
             try {
                 Utils.verifyOffMainThread();
 
-                StreamingDataRequest request = StreamingDataRequest.getRequestForVideoId(videoId);
-                if (request != null) {
-                    var stream = request.getStream();
+                var future = currentVideoStream;
+                if (future != null) {
+                    final long maxSecondsToWait = 20;
+                    var stream = future.get(maxSecondsToWait, TimeUnit.SECONDS);
                     if (stream != null) {
-                        Logger.printDebug(() -> "Overriding video stream: " + videoId);
+                        Logger.printDebug(() -> "Overriding video stream");
                         return stream;
                     }
-                }
 
-                Logger.printDebug(() -> "Not overriding streaming data (video stream is null): " + videoId);
-            } catch (Exception ex) {
+                    Logger.printDebug(() -> "Not overriding streaming data (video stream is null)");
+                }
+            } catch (TimeoutException ex) {
+                Logger.printInfo(() -> "getStreamingData timed out", ex);
+            } catch (InterruptedException ex) {
+                Logger.printException(() -> "getStreamingData interrupted", ex);
+                Thread.currentThread().interrupt(); // Restore interrupt status flag.
+            } catch (ExecutionException ex) {
                 Logger.printException(() -> "getStreamingData failure", ex);
             }
         }
