+++ stub/src/main/java/org/chromium/net/UrlRequest.java
@@ -1,4 +1,8 @@
 package org.chromium.net;
 
 public abstract class UrlRequest {
+    public abstract class Builder {
+        public abstract Builder addHeader(String name, String value);
+        public abstract UrlRequest build();
+    }
 }
